+++ Changelog.md
@@ -4,6 +4,27 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)//
 
+## 4.4.0-rc.15
+### Changed
+ - Updated @formio/core@2.4.0-rc.7
+ - Updated @formio/vm@1.0.0-rc.13
+ - Updated @formio/js@5.1.0-rc.10
+
+## 4.4.0-rc.14
+ - FIO-8428: Fixes save submission action data transformation
+ 
+## 4.4.0-rc.13
+ - FIO-9503: fixed nested form data not appearing in emails
+
+## 4.4.0-rc.12
+### Changed
+ - Updated @formio/core@2.4.0-rc.6
+ - Updated @formio/vm@1.0.0-rc.12
+ - Updated @formio/js@5.1.0-rc.9
+ - FIO-8644 fixed hidden calculated fields on update submission
+
+ - FIO-9499: Add test for email action with edit grid
+ 
 ## 4.4.0-rc.9
 ### Changed
  - Updated @formio/vm@1.0.0-rc.10